+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Implemented support for the Forge update JSON system (raoulvdberge)
 - Fixed crash in storage cache (raoulvdberge)
 - Fixed Crafting Pattern model (pauljoda)
+- Fixed Constructor not working on Botania flowers (raoulvdberge)
 
 ### 1.2.22
 - Fixed recipe for Processing Pattern Encoder not using oredictionary for the workbench (VT-14)